LayoutTests/http/tests/security/xssAuditor/script-tag-with-trailing-comment3-expected.txt

CONSOLE MESSAGE: line 6: Refused to execute a JavaScript script. Source code of script found within request.
